requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2
django-cors-headers==4.7.0
django-redis==5.4.0
django-resized==1.0.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
redis==6.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
